split integration into separate scrip
