+++ owiddata/generate-owiddata-stats.sh
@@ -9,18 +9,24 @@ OWID_COMMIT_JSON=$(curl -sS "https://api.github.com/repos/owid/covid-19-data/com
 export OWID_COMMIT_SHA=$(echo $OWID_COMMIT_JSON | python -c "import sys, json; print(json.load(sys.stdin)[0]['sha'])")
 export OWID_COMMIT_DATE=$(echo $OWID_COMMIT_JSON | python -c "import sys, json; print(json.load(sys.stdin)[0]['commit']['author']['date'])")
 
-# The output filenames
-VACCINE_PLATFORMS=owiddata/vaccine_platforms.csv
+# Location of JSON with stats
 OWID_STATS_JSON=owiddata/owiddata-stats.json
+
+# The temporary output filenames
+VACCINE_PLATFORMS=owiddata/vaccine_platforms.csv
+COUNTRY_BY_VAX=owiddata/country_by_vax.json
+
+# Locations for images
 OWID_MAP=owiddata/maps
 
 echo "Generating Our World in Data COVID-19 vaccine statistics"
-python owiddata/01.OWID.basicStats.py $OWID_STATS_JSON
+python owiddata/01.OWID.basicStats.py $OWID_STATS_JSON $COUNTRY_BY_VAX
 python owiddata/02.VIPER.basicStats.py $OWID_STATS_JSON  $VACCINE_PLATFORMS
+python owiddata/03.integrateDataSources.py $VACCINE_PLATFORMS $COUNTRY_BY_VAX
 #python owiddata/generate-owiddata-stats.py $OWID_STATS_JSON  $VACCINE_PLATFORMS $OWID_MAP
 
 # Clean up
-#rm $VACCINE_PLATFORMS
+#rm $VACCINE_PLATFORMS $COUNTRY_BY_VAX
 
 # After running this Python script to generate the figures, commit the figures
 # and run the version-figures.sh script to update the OWID_STATS_JSON with the
